packages/embedder/pyproject.toml
[project]
name = "embedder"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.76.0",
    "convex>=0.7.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
]
